Add unwind_block to StackPopAction, add comments
