+++ spec/lang/step/intrinsics.md
@@ -101,22 +101,6 @@ impl<M: Memory> Machine<M> {
 }
 ```
 
-Currently `Panic` carries no message and aborts directly.
-
-```rust
-impl<M: Memory> Machine<M> {
-    fn eval_intrinsic(
-        &mut self,
-        IntrinsicOp::Panic: IntrinsicOp,
-        arguments: List<(Value<M>, Type)>,
-        ret_ty: Type,
-    ) -> NdResult<Value<M>> {
-        // Stop machine immediatly without any additional checks.
-        throw_abort!("we panicked");
-    }
-}
-```
-
 `Abort` stopts the machine immediately.
 
 ```rust
@@ -127,7 +111,7 @@ impl<M: Memory> Machine<M> {
         arguments: List<(Value<M>, Type)>,
         ret_ty: Type,
     ) -> NdResult<Value<M>> {
-        throw_abort!("aborted");
+        throw_abort!();
     }
 }
 ```
