+++ spec/lang/syntax.md
@@ -362,11 +362,13 @@ pub enum Terminator {
         next_block: Option<BbName>,
         /// The block to jump to when this call unwinds.
         /// If `None`, UB will be raised when the function unwinds.
+        /// This comes with a well-formedness requirement: if the current block is a regular block,
+        /// `unwind_block` must be a cleanup block; otherwise, `unwind_block` must be a terminating block.
         unwind_block: Option<BbName>
     },
     /// Return from the current function.
     Return,
-    /// Starts unwinding, jump to the first cleanup block.
+    /// Starts unwinding, jump to the indicated cleanup block.
     StartUnwind(BbName),
     /// Ends this function call. The unwinding should continue at the caller's stack frame.
     ResumeUnwind,
@@ -406,7 +408,6 @@ pub enum IntrinsicOp {
     Abort,
     Assume,
     Exit,
-    Panic,
     PrintStdout,
     PrintStderr,
     Allocate,
@@ -485,8 +486,11 @@ pub struct BasicBlock {
 
 /// The kind of a basic block in the CFG.
 pub enum BbKind {
+    /// Regular blocks may use `Return` and `StartUnwind` but not `ResumeUnwind`.
     Regular,
+    /// Cleanup blocks may use `ResumeUnwind` but not `Return` or `StartUnwind`.
     Cleanup,
+    /// `Terminate` blocks may use neither `Return` nor `ResumeUnwind` nor `StartUnwind`.
     Terminate,
 }
