+++ spec/lang/well-formed.md
@@ -668,17 +668,18 @@ fn is_atomic_binop(op: IntBinOp) -> bool {
 }
 
 impl Terminator {
+    /// Check the terminator and the block transitions.
+    /// `block_kind` is the kind of the block in which the terminator is used.
     fn check_wf<T: Target>(
         self,
-        block: BasicBlock,
+        block_kind: BbKind,
         func: Function,
         prog: Program,
     ) -> Result<()> {
         use Terminator::*;
         match self {
             Goto(block_name) => {
-                ensure_wf(func.blocks.contains_key(block_name), "Terminator::Goto: next block does not exist")?;
-                ensure_wf(block.kind == func.get_block_kind(block_name), "Terminator::Goto: next block has the wrong block kind")?;
+               func.check_next_block(block_name, block_kind)?;
             }
             Switch { value, cases, fallback } => {
                 let ty = value.check_wf::<T>(func.locals, prog)?;
@@ -692,13 +693,11 @@ impl Terminator {
                 // Ensure the switch cases are all valid.
                 for (case, next_block) in cases.iter() {
                     ensure_wf(switch_ty.can_represent(case), "Terminator::Switch: value does not fit in switch type")?;
-                    ensure_wf(func.blocks.contains_key(next_block), "Terminator::Switch: next block does not exist")?;
-                    ensure_wf(block.kind == func.get_block_kind(next_block), std::format!("Terminator::Switch: next block in case {} has the wrong block kind", case).as_str())?;
+                    func.check_next_block(next_block, block_kind)?;
                 }
 
                 // Ensure the fallback is valid.
-                ensure_wf(func.blocks.contains_key(fallback), "Terminator::Switch: fallback block does not exist")?;
-                ensure_wf(block.kind == func.get_block_kind(fallback), "Terminator::Switch: fallback block has the wrong block kind")?;
+                func.check_next_block(fallback, block_kind)?;
             }
             Unreachable => {}
             Intrinsic { intrinsic, arguments, ret, next_block } => {
@@ -721,8 +720,7 @@ impl Terminator {
                 }
 
                 if let Some(next_block) = next_block {
-                    ensure_wf(func.blocks.contains_key(next_block), "Terminator::Intrinsic: next block does not exist")?;
-                    ensure_wf(block.kind == func.get_block_kind(next_block), "Terminator::Intrinsic: next block has the wrong block kind")?;
+                    func.check_next_block(next_block, block_kind)?;
                 }
             }
             Call { callee, calling_convention: _, arguments, ret, next_block, unwind_block } => {
@@ -738,32 +736,29 @@ impl Terminator {
                 }
 
                 if let Some(next_block) = next_block {
-                    ensure_wf(func.blocks.contains_key(next_block), "Terminator::Call: next block does not exist")?;
-                    ensure_wf(block.kind == func.get_block_kind(next_block), "Terminator::Call: next block has the wrong block kind")?;
+                    func.check_next_block(next_block, block_kind)?;
                 }
 
                 if let Some(unwind_block) = unwind_block {
-                    ensure_wf(func.blocks.contains_key(unwind_block), "Terminator::Call: unwind block does not exist")?;
-                    match block.kind{
-                            BbKind::Regular => {
-                                ensure_wf(func.get_block_kind(unwind_block) == BbKind::Cleanup, "Terminator::Call: unwind block has the wrong block kind")?;
-                            }
-                            _ => {
-                                ensure_wf(func.get_block_kind(unwind_block) == BbKind::Terminate, "Terminator::Call: unwind block has the wrong block kind")?;
-                            }
+                    match block_kind {
+                        BbKind::Regular => {
+                            func.check_next_block(unwind_block, BbKind::Cleanup)?;
                         }
+                        _ => {
+                            func.check_next_block(unwind_block, BbKind::Terminate)?;
+                        }
+                    }
                 }
             }
             Return => {
-                 ensure_wf(block.kind == BbKind::Regular, "Terminator::Return has to be called in a regular block")?;
+                 ensure_wf(block_kind == BbKind::Regular, "Terminator::Return has to be called in a regular block")?;
             }
             StartUnwind(unwind_block) => {
-                ensure_wf(func.blocks.contains_key(unwind_block), "Terminator::StartUnwind: unwind block does not exist")?;
-                ensure_wf(block.kind == BbKind::Regular, "Terminator::StartUnwind has to be called in regular block")?;
-                ensure_wf(func.get_block_kind(unwind_block) == BbKind::Cleanup, "Terminator::StartUnwind: next block has the wrong block kind")?;
+                ensure_wf(block_kind == BbKind::Regular, "Terminator::StartUnwind has to be called in regular block")?;
+                func.check_next_block(unwind_block, BbKind::Cleanup)?;
             }
             ResumeUnwind => {
-                 ensure_wf(block.kind == BbKind::Cleanup, "Terminator::ResumeUnwind: has to be called in cleanup block")?;
+                 ensure_wf(block_kind == BbKind::Cleanup, "Terminator::ResumeUnwind: has to be called in cleanup block")?;
             }
         }
 
@@ -798,14 +793,21 @@ impl Function {
             for statement in block.statements {
                 statement.check_wf::<T>(self, prog)?;
             }
-            block.terminator.check_wf::<T>(block, self, prog)?;
+            block.terminator.check_wf::<T>(block.kind, self, prog)?;
         }
 
         ret(())
     }
 
-    fn get_block_kind(self, block_name: BbName) -> BbKind {
-        self.blocks.get(block_name).expect("get_block_kind: Block does not exist").kind
+    /// Checks whether the next block exists and has the correct block kind.
+    fn check_next_block(self, next_block_name: BbName, expected_block_kind: BbKind) -> Result <()> {
+        if let Some(next_block) = self.blocks.get(next_block_name){
+            ensure_wf(next_block.kind == expected_block_kind, "Terminator: next block has the wrong block kind")?;
+        }
+        else{
+            throw_ill_formed!("Terminator: next block does not exist");
+        }
+        ret(())
     }
 }
