+++ spec/prelude/main.md
@@ -24,7 +24,7 @@ pub enum TerminationInfo {
     /// The program was executed and the machine stopped without error.
     MachineStop,
     /// The program terminated with a panic
-    Abort(String),
+    Abort,
     /// The program was ill-formed.
     IllFormed(String),
     /// The program did not terminate but no thread can make progress.
@@ -46,8 +46,8 @@ macro_rules! throw_ub {
     };
 }
 macro_rules! throw_abort {
-    ($($tt:tt)*) => {
-        do yeet TerminationInfo::Abort(format!($($tt)*))
+    () => {
+        do yeet TerminationInfo::Abort
     }
 }
 macro_rules! throw_machine_stop {
