+++ tooling/minimize/tests/panic/array_out_of_bounds.stderr
@@ -1,1 +1,1 @@
-fatal error: Panic: we panicked
+fatal error: program aborted
