+++ tooling/minimize/tests/panic/assert.stderr
@@ -1,1 +1,1 @@
-fatal error: Panic: we panicked
+fatal error: program aborted
