+++ tooling/minitest/src/tests/call.rs
@@ -131,43 +131,3 @@ fn ret_mismatch() {
     dump_program(p);
     assert_ub::<BasicMem>(p, "call ABI violation: return types are not compatible");
 }
-
-#[test]
-fn next_block_non_exist() {
-    let locals = [<()>::get_type()];
-
-    let b0 = block!(storage_live(0), Terminator::Call {
-        callee: fn_ptr_internal(1),
-        calling_convention: CallingConvention::C,
-        arguments: list![by_value(unit())],
-        ret: local(0),
-        next_block: Some(BbName(Name::from_internal(2))),
-        unwind_block: Some(BbName(Name::from_internal(1))),
-    });
-    let b1 = block!(exit());
-
-    let f = function(Ret::No, 0, &locals, &[b0, b1]);
-    let p = program(&[f, other_f()]);
-    dump_program(p);
-    assert_ill_formed::<BasicMem>(p, "Terminator::Call: next block does not exist");
-}
-
-#[test]
-fn unwind_block_non_exist() {
-    let locals = [<()>::get_type()];
-
-    let b0 = block!(storage_live(0), Terminator::Call {
-        callee: fn_ptr_internal(1),
-        calling_convention: CallingConvention::C,
-        arguments: list![by_value(unit())],
-        ret: local(0),
-        next_block: Some(BbName(Name::from_internal(1))),
-        unwind_block: Some(BbName(Name::from_internal(2))),
-    });
-    let b1 = block!(exit());
-
-    let f = function(Ret::No, 0, &locals, &[b0, b1]);
-    let p = program(&[f, other_f()]);
-    dump_program(p);
-    assert_ill_formed::<BasicMem>(p, "Terminator::Call: unwind block does not exist");
-}
