+++ tooling/minitest/src/tests/unwinding.rs
@@ -1,11 +1,43 @@
 use crate::*;
 
+/// Test if the execution reaches the terminate block. The test calls a function that panics, and then calls it again in the cleanup block.
+#[test]
+fn reach_terminate_block() {
+    let mut p = ProgramBuilder::new();
+
+    let panic_fn = {
+        let mut f = p.declare_function();
+        let resume = f.cleanup_block(|f| f.resume_unwind());
+        f.start_unwind(resume);
+        p.finish_function(f)
+    };
+
+    let main_fn = {
+        let mut f = p.declare_function();
+        let terminate = f.terminating_block(|f| {
+            f.exit();
+        });
+        let cleanup = f.cleanup_block(|f| {
+            f.call(unit_place(), fn_ptr(panic_fn), &[], terminate);
+            f.unreachable();
+        });
+
+        f.call(unit_place(), fn_ptr(panic_fn), &[], cleanup);
+        f.unreachable();
+        p.finish_function(f)
+    };
+    let p = p.finish_program(main_fn);
+    dump_program(p);
+    assert_stop::<BasicMem>(p);
+}
+
+/// Test a program that starts unwinding and aborts in the cleanup block.
 #[test]
 fn abort_in_cleanup() {
     let mut p = ProgramBuilder::new();
     let mut f = p.declare_function();
 
-    let cleanup = f.cleanup(|f| {
+    let cleanup = f.cleanup_block(|f| {
         f.abort();
     });
 
@@ -13,17 +45,19 @@ fn abort_in_cleanup() {
     let f = p.finish_function(f);
     let p = p.finish_program(f);
     dump_program(p);
-    assert_abort::<BasicMem>(p, "aborted");
+    assert_abort::<BasicMem>(p);
 }
 
+/// This test calls `print` in the cleanup block and checks whether the value is actually printed.
 #[test]
 fn start_unwind_in_main() {
     let mut p = ProgramBuilder::new();
     let mut f = p.declare_function();
 
-    let cleanup = f.cleanup(|f| {
+    let cleanup = f.cleanup_block(|f| {
         f.print(const_int(42));
-        f.exit(); //Call exit() instead of abort, because there is currently no way to access the standard output if the program has aborted.
+        // Call `exit()` instead of `abort()`, because there is currently no way to access the standard output if the program has aborted.
+        f.exit();
     });
     f.start_unwind(cleanup);
     let f = p.finish_function(f);
@@ -32,6 +66,23 @@ fn start_unwind_in_main() {
     assert_eq!(get_stdout::<BasicMem>(p).unwrap(), &["42"]);
 }
 
+/// This test calls a recursive function that will panic after some recursive calls.
+/// The function is structured as follows:
+/// ```
+/// fn recursive_fn(arg: i32){  
+///     print(arg);  
+///     if arg == 0 {  
+///         StartUnwind();  
+///     }  
+///     else{  
+///         recursive_fn(arg-1);  
+///     }  
+///     
+///     --Cleanup-- {  
+///         print(arg);  
+///     }  
+/// }
+/// ```
 #[test]
 fn unwind_recursive_func() {
     let mut p = ProgramBuilder::new();
@@ -42,11 +93,12 @@ fn unwind_recursive_func() {
         let var = f.declare_local::<i32>();
         let ret = f.declare_ret::<i32>();
 
-        let cleanup_resume = f.cleanup_resume();
-        let cleanup_print = f.cleanup(|f| {
+        let cleanup_resume = f.cleanup_block(|f| f.resume_unwind());
+        let cleanup_print = f.cleanup_block(|f| {
             f.print(load(arg));
             f.goto(cleanup_resume);
         });
+        f.print(load(arg));
         f.if_(
             eq(load(arg), const_int(0)),
             |f| f.start_unwind(cleanup_resume),
@@ -65,12 +117,12 @@ fn unwind_recursive_func() {
         let mut main_fn = p.declare_function();
         let var = main_fn.declare_local::<i32>();
 
-        let cleanup = main_fn.cleanup(|f| {
+        let cleanup = main_fn.cleanup_block(|f| {
             f.exit();
         });
 
         main_fn.storage_live(var);
-        main_fn.call(var, fn_ptr(f), &[by_value(const_int(10))], cleanup);
+        main_fn.call(var, fn_ptr(f), &[by_value(const_int(6))], cleanup);
         main_fn.storage_dead(var);
         main_fn.exit();
         p.finish_function(main_fn)
@@ -79,10 +131,11 @@ fn unwind_recursive_func() {
     let p = p.finish_program(main_fn);
     dump_program(p);
     assert_eq!(get_stdout::<BasicMem>(p).unwrap(), &[
-        "1", "2", "3", "4", "5", "6", "7", "8", "9", "10"
+        "6", "5", "4", "3", "2", "1", "0", "1", "2", "3", "4", "5", "6",
     ]);
 }
 
+/// A test case with non-terminator statements in the cleanup block.
 #[test]
 fn statements_in_cleanup() {
     let mut p = ProgramBuilder::new();
@@ -92,7 +145,7 @@ fn statements_in_cleanup() {
         let arg = f.declare_arg::<i32>();
         let var = f.declare_local::<i32>();
         let _ret = f.declare_ret::<i32>();
-        let cleanup = f.cleanup(|f| {
+        let cleanup = f.cleanup_block(|f| {
             f.assign(var, add_unchecked(load(arg), const_int(3)));
             f.assign(var, mul_unchecked(load(var), load(arg)));
             f.assign(var, shl_unchecked(load(var), const_int(2)));
@@ -108,7 +161,7 @@ fn statements_in_cleanup() {
         let mut main_fn = p.declare_function();
         let var = main_fn.declare_local::<i32>();
 
-        let cleanup = main_fn.cleanup(|f| {
+        let cleanup = main_fn.cleanup_block(|f| {
             f.exit();
         });
 
@@ -123,20 +176,22 @@ fn statements_in_cleanup() {
     assert_eq!(get_stdout::<BasicMem>(p).unwrap(), &["520"]);
 }
 
+/// This test prints before and after a function call.
+/// The called function panics, so only the first print statement should be executed.
 #[test]
 fn print_after_unwind() {
     let mut p = ProgramBuilder::new();
 
     let f = {
         let mut f = p.declare_function();
-        let cleanup = f.cleanup_resume();
+        let cleanup = f.cleanup_block(|f| f.resume_unwind());
         f.start_unwind(cleanup);
         p.finish_function(f)
     };
 
     let main_fn = {
         let mut main_fn = p.declare_function();
-        let cleanup = main_fn.cleanup(|f| {
+        let cleanup = main_fn.cleanup_block(|f| {
             f.exit();
         });
 
@@ -151,18 +206,48 @@ fn print_after_unwind() {
     assert_eq!(get_stdout::<BasicMem>(p).unwrap(), &["1"]);
 }
 
+/// This test resumes unwinding in the start function, which should result in UB.
 #[test]
 fn resume_in_main() {
     let mut p = ProgramBuilder::new();
     let mut f = p.declare_function();
-    let cleanup = f.cleanup_resume();
+    let cleanup = f.cleanup_block(|f| f.resume_unwind());
     f.start_unwind(cleanup);
     let f = p.finish_function(f);
     let p = p.finish_program(f);
     dump_program(p);
-    assert_ub::<BasicMem>(p, "the start function must not resume");
+    assert_ub::<BasicMem>(p, "the function at the bottom of the stack must not unwind");
 }
 
+/// This test creates a new thread and resumes unwinding at the bottom of the stack, which should result in UB.
+#[test]
+fn resume_in_thread() {
+    let mut p = ProgramBuilder::new();
+
+    let f = {
+        let mut f = p.declare_function();
+        let _ = f.declare_arg::<*const ()>();
+        let cleanup = f.cleanup_block(|f| f.resume_unwind());
+        f.start_unwind(cleanup);
+        p.finish_function(f)
+    };
+
+    let main_fn = {
+        let mut main_fn = p.declare_function();
+        let x = main_fn.declare_local::<i32>();
+        main_fn.storage_live(x);
+        main_fn.spawn(f, null(), x);
+        main_fn.join(load(x));
+        main_fn.exit();
+        p.finish_function(main_fn)
+    };
+
+    let p = p.finish_program(main_fn);
+    dump_program(p);
+    assert_ub::<BasicMem>(p, "the function at the bottom of the stack must not unwind");
+}
+
+/// This test resumes unwinding, but no `unwind_block` is specified, which should result in UB.
 #[test]
 fn resume_no_unwind_block() {
     let mut p = ProgramBuilder::new();
@@ -176,7 +261,7 @@ fn resume_no_unwind_block() {
         f.if_(
             eq(load(var), const_int(0)),
             |f| {
-                let cleanup = f.cleanup_resume();
+                let cleanup = f.cleanup_block(|f| f.resume_unwind());
                 f.start_unwind(cleanup);
             },
             |f| {
@@ -190,7 +275,7 @@ fn resume_no_unwind_block() {
     let main_fn = {
         let mut main_fn = p.declare_function();
 
-        let cleanup = main_fn.cleanup(|f| {
+        let cleanup = main_fn.cleanup_block(|f| {
             f.abort();
         });
         main_fn.call(unit_place(), fn_ptr(f), &[by_value(const_int(3))], cleanup);
@@ -200,5 +285,8 @@ fn resume_no_unwind_block() {
     };
     let p = p.finish_program(main_fn);
     dump_program(p);
-    assert_ub::<BasicMem>(p, "unwinding from a function where caller did not specify unwind_block");
+    assert_ub::<BasicMem>(
+        p,
+        "unwinding from a function where caller did not specify an unwind_block",
+    );
 }
