+++ tooling/miniutil/src/fmt/function.rs
@@ -231,7 +231,6 @@ fn fmt_terminator(t: Terminator, comptypes: &mut Vec<CompType>) -> String {
                 IntrinsicOp::Abort => "abort",
                 IntrinsicOp::Assume => "assume",
                 IntrinsicOp::Exit => "exit",
-                IntrinsicOp::Panic => "panic",
                 IntrinsicOp::PrintStdout => "print",
                 IntrinsicOp::PrintStderr => "eprint",
                 IntrinsicOp::Allocate => "allocate",
